dakldjlakdknajhd djahkdwjhawd

+++ resources/js/components/admin/Logout.vue
@@ -7,11 +7,33 @@
 <script>
 	export default {
 		mounted(){
-			axios.post('/logout')
-			.then(res=>{
-				window.location='/login'
-				sessionStorage.clear()
-			})
+
+			this.$swal({
+            title: 'Are you sure?',
+            text: "You want to Logout!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              this.$swal({
+                  icon: 'success',
+                  title: 'Logout Success',
+                  showConfirmButton: false,
+                  timer: 1500
+                })
+              axios.post('/logout')
+      				.then(res=>{
+      					window.location='/login'
+      					sessionStorage.clear()
+      				})
+            }else{
+            	this.$router.go(-1)
+            }
+          })
+			
 		}
 	}
 </script>
